websitee/python_decoder/requirements.txt
fastapi==0.115.2
pydantic==2.8.2
uvicorn==0.30.6
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.0.1



